Add files via upload
You can select an algorithm and type in an array of integers to observe the implementation in main.cpp.
sort.cpp contains the class of the entire process of sorting. 
All algorithms are considered member functions of the class, and the function Option_Initialize_Sort will initialize the entire process (input, driver, print...etc.).